formularios acabados

+++ src/app/components/courses/listar-courses/listar-courses.component.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { Course } from '../../../models/course';
 import { CourseService } from '../../../services/course.service';
 import { RouterLink, RouterModule } from '@angular/router';
+import { Student } from '../../../models/student';
+import { StudentService } from '../../../services/student.service';
 
 @Component({
   selector: 'app-listar-courses',
@@ -11,13 +13,21 @@ import { RouterLink, RouterModule } from '@angular/router';
 })
 export class ListarCoursesComponent {
   courses: Course[] = [];
+  students: Student[] = [];
 
-  constructor(private courseService: CourseService) {}
+  constructor(
+    private courseService: CourseService,
+    private studentService: StudentService
+  ) {}
 
   ngOnInit(): void {
     this.courseService.getCourses().subscribe((courses) => {
       this.courses = courses;
       console.log(courses);
+    });
+    this.studentService.getStudents().subscribe((students) => {
+      this.students = students;
+      console.log(students);
     });
   }
 
@@ -27,17 +37,19 @@ export class ListarCoursesComponent {
     });
   }
 
-  getCourse(id:string){
+  getCourse(id: string) {
     this.courseService.getCouse(id).subscribe((course) => {
       console.log(course);
     });
   }
 
-  eliminarCourse(id: string){
-    this.courseService.deleteCourse(id).subscribe(() => {
-      this.getCoursesAll();
-    });
+  eliminarCourse(id: string) {
+    if (this.students.find((s) => s.course === id)) {
+      this.courseService.deleteCourse(id).subscribe(() => {
+        this.getCoursesAll();
+      });
+    } else {
+      alert('Este Curso tiene alumnos asociados');
+    }
   }
-
-
 }
